maximum length defaults to the cutoff radius

+++ src/applications/dynamicProps/CorrelationFunction.hpp
@@ -96,7 +96,7 @@ class CorrelationFunction {
         int nTimeBins_;
         std::vector<double> histogram_;
         std::vector<int> count_;
-        std::vector<int> time_;
+        std::vector<double> time_;
         
         SimInfo* info_;
         int storageLayout_;
